backend/requirements.txt
Flask==2.2.5
flask-cors==4.0.0
flask-mysqldb==2.0.0
Werkzeug==2.2.3
